# server/go/model_student.go
/*
 * ITU REST API
 *
 * This is documentation for our REST api for mandatory exercise 1 in the distributed systems course
 *
 * API version: 1.0.1½
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

import (
	"encoding/json"
	"io"
	"io/ioutil"
)

type Student struct {
	Id int64 `json:"id"`

	Name string `json:"name"`
}

func parseStudent(body io.ReadCloser) Student {
	resp, _ := ioutil.ReadAll(body)

	var student Student
	json.Unmarshal(resp, &student)
	return student
}